llmserver/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via
    #   langchain
    #   langchain-community
aiosignal==1.3.1
    # via aiohttp
annotated-types==0.6.0
    # via pydantic
anyio==4.3.0
    # via
    #   httpx
    #   langchain-core
    #   openai
    #   sse-starlette
    #   starlette
attrs==23.2.0
    # via aiohttp
cachetools==5.3.3
    # via google-auth
certifi==2024.2.2
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via uvicorn
dataclasses-json==0.6.4
    # via
    #   langchain
    #   langchain-community
distro==1.9.0
    # via openai
fastapi==0.110.0
    # via
    #   -r requirements.in
    #   sse-starlette
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
google-api-core[grpc]==2.17.1
    # via
    #   google-cloud-core
    #   google-cloud-firestore
google-auth==2.28.2
    # via
    #   google-api-core
    #   google-cloud-core
google-cloud-core==2.4.1
    # via google-cloud-firestore
google-cloud-firestore==2.15.0
    # via langchain-google-firestore
googleapis-common-protos==1.63.0
    # via
    #   google-api-core
    #   grpcio-status
greenlet==3.0.3
    # via sqlalchemy
grpcio==1.62.1
    # via
    #   google-api-core
    #   grpcio-status
grpcio-status==1.62.1
    # via google-api-core
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.4
    # via httpx
httpx==0.27.0
    # via
    #   langserve
    #   openai
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
langchain==0.1.9
    # via
    #   -r requirements.in
    #   langserve
langchain-community==0.0.24
    # via
    #   langchain
    #   langchain-google-firestore
langchain-core==0.1.27
    # via
    #   -r requirements.in
    #   langchain
    #   langchain-community
    #   langchain-google-firestore
    #   langchain-openai
langchain-google-firestore==0.1.1
    # via -r requirements.in
langchain-openai==0.0.8
    # via -r requirements.in
langserve==0.0.51
    # via -r requirements.in
langsmith==0.1.10
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
marshmallow==3.21.0
    # via dataclasses-json
more-itertools==10.2.0
    # via langchain-google-firestore
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.4
    # via
    #   langchain
    #   langchain-community
openai==1.13.3
    # via langchain-openai
orjson==3.9.15
    # via
    #   langserve
    #   langsmith
packaging==23.2
    # via
    #   langchain-core
    #   marshmallow
proto-plus==1.23.0
    # via google-cloud-firestore
protobuf==4.25.3
    # via
    #   -r requirements.in
    #   google-api-core
    #   google-cloud-firestore
    #   googleapis-common-protos
    #   grpcio-status
    #   proto-plus
pyasn1==0.5.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.3.0
    # via google-auth
pydantic==2.6.4
    # via
    #   -r requirements.in
    #   fastapi
    #   langchain
    #   langchain-core
    #   langserve
    #   langsmith
    #   openai
pydantic-core==2.16.3
    # via pydantic
python-dotenv==1.0.1
    # via -r requirements.in
pyyaml==6.0.1
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
regex==2023.12.25
    # via tiktoken
requests==2.31.0
    # via
    #   google-api-core
    #   langchain
    #   langchain-community
    #   langchain-core
    #   langsmith
    #   tiktoken
rsa==4.9
    # via google-auth
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
    #   openai
sqlalchemy==2.0.27
    # via
    #   langchain
    #   langchain-community
sse-starlette==1.8.2
    # via -r requirements.in
starlette==0.36.3
    # via
    #   fastapi
    #   sse-starlette
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
tiktoken==0.6.0
    # via langchain-openai
tqdm==4.66.2
    # via openai
typing-extensions==4.10.0
    # via
    #   fastapi
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
urllib3==2.2.1
    # via requests
uvicorn==0.28.0
    # via
    #   -r requirements.in
    #   sse-starlette
yarl==1.9.4
    # via aiohttp
